Move classes relevant to unification and transforming conceptmaps down into the logic package
