+++ logic/concludable/ConjunctionConcludable.java
@@ -21,7 +21,6 @@ import grakn.common.collection.Pair;
 import grakn.core.common.exception.GraknException;
 import grakn.core.common.parameters.Label;
 import grakn.core.logic.Rule;
-import grakn.core.logic.Unification;
 import grakn.core.pattern.Conjunction;
 import grakn.core.pattern.constraint.Constraint;
 import grakn.core.pattern.constraint.thing.HasConstraint;
@@ -31,6 +30,7 @@ import grakn.core.pattern.constraint.thing.ThingConstraint;
 import grakn.core.pattern.constraint.thing.ValueConstraint;
 import grakn.core.pattern.equivalence.AlphaEquivalence;
 import grakn.core.pattern.variable.Variable;
+import grakn.core.logic.transform.Unifier;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -52,13 +52,18 @@ public abstract class ConjunctionConcludable<CONSTRAINT extends Constraint, U ex
         return new Extractor(conjunction.variables()).concludables();
     }
 
-    public Stream<Pair<Rule, Unification>> findUnifiableRules(Stream<Rule> allRules) {
+    public Stream<Pair<Rule, Unifier>> findUnifiableRules(Stream<Rule> allRules) {
+        // TODO Get rules internally
         return allRules.flatMap(rule -> rule.possibleThenConcludables().stream()
                                                .flatMap(this::unify).map(unifiedBase -> new Pair<>(rule, unifiedBase))
         );
     }
 
-    private Stream<Unification> unify(ThenConcludable<?, ?> unifyWith) {
+    public Stream<Unifier> applicableRuleUnifiers() {
+        return Stream.empty(); // TODO This is the signiature we really want for findUnifiableRules
+    }
+
+    private Stream<Unifier> unify(ThenConcludable<?, ?> unifyWith) {
         if (unifyWith instanceof ThenConcludable.Relation) return unify((ThenConcludable.Relation) unifyWith);
         else if (unifyWith instanceof ThenConcludable.Has) return unify((ThenConcludable.Has) unifyWith);
         else if (unifyWith instanceof ThenConcludable.Isa) return unify((ThenConcludable.Isa) unifyWith);
@@ -66,19 +71,19 @@ public abstract class ConjunctionConcludable<CONSTRAINT extends Constraint, U ex
         else throw GraknException.of(ILLEGAL_STATE);
     }
 
-    Stream<Unification> unify(ThenConcludable.Relation unifyWith) {
+    Stream<Unifier> unify(ThenConcludable.Relation unifyWith) {
         return Stream.empty();
     }
 
-    Stream<Unification> unify(ThenConcludable.Has unifyWith) {
+    Stream<Unifier> unify(ThenConcludable.Has unifyWith) {
         return Stream.empty();
     }
 
-    Stream<Unification> unify(ThenConcludable.Isa unifyWith) {
+    Stream<Unifier> unify(ThenConcludable.Isa unifyWith) {
         return Stream.empty();
     }
 
-    Stream<Unification> unify(ThenConcludable.Value unifyWith) {
+    Stream<Unifier> unify(ThenConcludable.Value unifyWith) {
         return Stream.empty();
     }
 
@@ -149,13 +154,13 @@ public abstract class ConjunctionConcludable<CONSTRAINT extends Constraint, U ex
         }
 
         @Override
-        public Stream<Unification> unify(ThenConcludable.Relation unifyWith) {
+        public Stream<Unifier> unify(ThenConcludable.Relation unifyWith) {
             // Check the relation variables' isa constraint labels and prune if there is no intersection
 
             // Find all roleplayer mapping combinations, which should have the form:
             // Set<List<Pair<RelationConstraint.RolePlayer, RelationConstraint.RolePlayer>>> rolePlayerMappings
             // For each, prune if there is no label intersection (or any other pruning, e.g. by value)
-            // Then build a Unification for each valid combination
+            // Then build a Unifier for each valid combination
             return Stream.empty(); // TODO
         }
 
@@ -182,7 +187,7 @@ public abstract class ConjunctionConcludable<CONSTRAINT extends Constraint, U ex
         }
 
         @Override
-        public Stream<Unification> unify(ThenConcludable.Has unifyWith) {
+        public Stream<Unifier> unify(ThenConcludable.Has unifyWith) {
             return null;
         }
 
@@ -209,11 +214,11 @@ public abstract class ConjunctionConcludable<CONSTRAINT extends Constraint, U ex
         }
 
         @Override
-        public Stream<Unification> unify(ThenConcludable.Isa headIsa) {
+        public Stream<Unifier> unify(ThenConcludable.Isa headIsa) {
             Set<Label> possibleLabels = constraint().getTypeHints();
             Set<Label> otherLabels = headIsa.constraint().getTypeHints();
             possibleLabels.retainAll(otherLabels);
-            return Stream.of(new Unification(this, headIsa, null)); // TODO Add variable mapping
+            return Stream.of(new Unifier(this, headIsa, null)); // TODO Add variable mapping
         }
 
         @Override
@@ -239,7 +244,7 @@ public abstract class ConjunctionConcludable<CONSTRAINT extends Constraint, U ex
         }
 
         @Override
-        public Stream<Unification> unify(ThenConcludable.Value unifyWith) {
+        public Stream<Unifier> unify(ThenConcludable.Value unifyWith) {
             return null;
         }
