+++ logic/transform/VariableMapper.java
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package grakn.core.reasoner.resolution;
+package grakn.core.logic.transform;
 
 import grakn.core.concept.answer.ConceptMap;
 import grakn.core.logic.concludable.ConjunctionConcludable;
@@ -26,7 +26,7 @@ import java.util.Map;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
-public class VariableMapper implements ConceptMapTransformer{
+public class VariableMapper implements ConceptMapTransformer {
 
     private Map<Variable, Variable> map;
