+++ reasoner/resolution/resolver/ConjunctionResolver.java
@@ -27,8 +27,8 @@ import grakn.core.pattern.Conjunction;
 import grakn.core.reasoner.resolution.MockTransaction;
 import grakn.core.reasoner.resolution.ResolutionRecorder;
 import grakn.core.reasoner.resolution.ResolverRegistry;
-import grakn.core.reasoner.resolution.TransformedConceptMap;
-import grakn.core.reasoner.resolution.VariableMapper;
+import grakn.core.logic.transform.TransformedConceptMap;
+import grakn.core.logic.transform.VariableMapper;
 import grakn.core.reasoner.resolution.framework.Request;
 import grakn.core.reasoner.resolution.framework.ResolutionAnswer;
 import grakn.core.reasoner.resolution.framework.Resolver;
